requirements.txt

openai>=1.0.0
playwright>=1.40.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.30.0
colorama>=0.4.6
pandas>=2.0.0
#shlex
google-generativeai>=0.3.0
google-genai>=0.1.0
fake-useragent>=1.5.0


